requirements.txt
asgiref==3.2.10
Django==3.0.8
django-multiselectfield==0.1.12
django-request==1.5.5
gunicorn==20.0.4
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
